insertExtensionBundlesConfig into own module
