+++ build-system/tasks/extension-generator/bento/index.js
@@ -25,8 +25,10 @@ const argv = require('minimist')(process.argv.slice(2));
 const fs = require('fs').promises;
 const log = require('fancy-log');
 const path = require('path');
+const {
+  insertExtensionBundlesConfig,
+} = require('../insert-extension-bundles-config');
 const {cyan, green, red, yellow} = require('ansi-colors');
-const {insertExtensionBundlesConfig} = require('../');
 
 const EXTENSIONS_DIR = path.join(__dirname, '../../../../extensions');
 const TEMPLATE_DIR = path.join(__dirname, 'amp-__component_name_hyphenated__');
